+++ R/ympes-package.R
@@ -1,0 +1,9 @@
+#' @keywords internal
+"_PACKAGE"
+
+## usethis namespace: start
+## usethis namespace: end
+NULL
+
+#' @useDynLib ympes, .registration = TRUE
+NULL